+++ sql_template/l1/fact_ths_daily.sql
@@ -11,20 +11,20 @@ create table if not exists l1.fact_ths_daily (
     ths_type        string      comment 'I行业/N概念指数/S特色指数',
     ths_type_str    string      comment '概念/行业/地域/特色',
 
-    `close`         float       comment '收盘点位',
-    `open`          float       comment '开盘点位',
-    `high`          float       comment '最高点位',
-    `low`           float       comment '最低点位',
-    `pre_close`     float       comment '昨日收盘点',
-    `avg_price`     float       comment '平均点位',
-    `change`        float       comment '涨跌点位',
-    `change_pct`    float       comment '涨跌幅',
-    `vol`           float       comment '成交量',
-    `turnover_rate` float       comment '换手率',
-    `total_mv`      float       comment '总市值',
-    `float_mv`      float       comment '流通市值',
-    `pe_ttm`        float       comment 'PE TTM',
-    `pb_mrq`        float       comment 'PB MRQ'
+    `close`         double       comment '收盘点位',
+    `open`          double       comment '开盘点位',
+    `high`          double       comment '最高点位',
+    `low`           double       comment '最低点位',
+    `pre_close`     double       comment '昨日收盘点',
+    `avg_price`     double       comment '平均点位',
+    `change`        double       comment '涨跌点位',
+    `change_pct`    double       comment '涨跌幅',
+    `vol`           double       comment '成交量',
+    `turnover_rate` double       comment '换手率',
+    `total_mv`      double       comment '总市值',
+    `float_mv`      double       comment '流通市值',
+    `pe_ttm`        double       comment 'PE TTM',
+    `pb_mrq`        double       comment 'PB MRQ'
 
 )  comment '同花顺概念行情'
 stored as orc;
