+++ sql_template/ods/money_flow_hsgt.sql
@@ -6,12 +6,12 @@
 -- ddl
 create table if not exists ods_incr.${table_name} (
     trade_date          string      comment '交易日期',
-    ggt_ss              float       comment '港股通（上海）',
-    ggt_sz              float       comment '港股通（深圳）',
-    hgt                 float       comment '沪股通',
-    sgt                 float       comment '深股通',
-    north_money         float       comment '北向资金',
-    south_money         float       comment '南向资金'
+    ggt_ss              double       comment '港股通（上海）',
+    ggt_sz              double       comment '港股通（深圳）',
+    hgt                 double       comment '沪股通',
+    sgt                 double       comment '深股通',
+    north_money         double       comment '北向资金',
+    south_money         double       comment '南向资金'
 )  comment '沪股通、深股通、港股通每日资金流向数据'
 partitioned by (
     pt_dt   string  comment '分区时间yyyy-mm-dd'
